starter-chat-agents/claude-code-reviewer/requirements.txt
streamlit>=1.28.0
anthropic>=0.7.0
python-dotenv>=1.0.0
pygments>=2.16.0
